mdcheck: reset sync_action to "idle" when stopped
When the mdcheck script stops because the pre-set duration is exceeded, it
will also set the sync action in the kernel to "idle". But when it is
stopped by a signal (e.g. when the systemd service running it is stopped),
it doesn't.  This is inconsistent behavior.

Move the code that switches the sync_action to "idle" into a cleanup
function that is always executed on exit. This requires separate "trap"
statements for EXIT(0) and signals, because otherwise a race condition may
arise between the cleanup code and the script body.

Signed-off-by: Martin Wilck <mwilck@suse.com>
(cherry picked from commit 8f0bcf062c45dacda4f889e8a6d99390b94068b6)